+++ CHANGELOG.md
@@ -26,6 +26,9 @@ This can also be enabled programmatically with `warnings.simplefilter('default',
 
 ### Deprecated
 
+## [2.7.10] - 2024-05-18
+### Added
+* support to optionally set `wrapmode` in templates (default `"WORD"` can optionally be set to `"CHAR"` to support wrapping on characters for scripts like Chinese or Japanese) - _cf._ [#1159](https://github.com/py-pdf/fpdf2/issues/1159)
 
 ## [2.7.9] - 2024-05-17
 ### Added